refactor: move set-content-type function
