+++ packages/verified-fetch/src/utils/type-guards.ts
@@ -1,0 +1,3 @@
+export function isPromise <T> (p?: any): p is Promise<T> {
+  return p?.then != null
+}
